requirements.txt
python-dotenv
langchain-openai
chainlit